refactor(server): extract estate_queries.go
Move estate snapshot, server/cluster/group snapshots, scoped
snapshot builder, and connection context.

Part of #75.
